iot2050setup: hardware-reset the cp210x device
For advanced board,hardware reset the cp210x to make the serial configuration take effect
righ away.

There is a design bug needing to take power cycle when siwtching the serial mode.
it already has fixed on the PG2 board,so don't need to prompt the user to reboot the the board
to make the change take effect.

Signed-off-by: chao zeng <chao.zeng@siemens.com>